updated documentation and vision constants refactoring [AK]
