+++ src/main/java/frc/lib/BobcatLib/CANdle/CANdleIOCANdle.java
@@ -26,17 +26,8 @@ public class CANdleIOCANdle implements CANdleIO {
         currState = state;
         
         switch (state) {
-            case INTAKING: //fire animation
-                leds.animate(BobcatUtil.getBuiltInAnimation(BuiltInAnimations.ColorFlow));
-                break;
-            case INTOOK: //green
-                leds.animate(new StrobeAnimation(0, 255, 0,  0, 1, CANdleConstants.LedCount)); 
-                break;
             case INTAKESTALL: //red strobe animation
                 leds.animate(new StrobeAnimation(255, 0, 0,  0, 0.25, CANdleConstants.LedCount));
-                break;
-            case NOTEHUNTING: 
-                leds.animate(BobcatUtil.getBuiltInAnimation(BuiltInAnimations.Rainbow));
                 break;
             case RESETPOSE: //strobe gold
                 leds.animate(new StrobeAnimation(255, 170, 0,  0, 0.25, CANdleConstants.LedCount));
@@ -46,9 +37,6 @@ public class CANdleIOCANdle implements CANdleIO {
                 break;
             case OUTAKE: //strobe red
                 leds.animate(new StrobeAnimation(255, 0, 0, 0, 0.75, CANdleConstants.LedCount));
-                break;
-            case FEED: //strobe green
-                leds.animate(new StrobeAnimation(0, 0, 255, 0, 0.75, CANdleConstants.LedCount));
                 break;
             case ALIGNING: // strobe white
                 leds.animate(new StrobeAnimation(255, 255, 255, 255, 0.75, CANdleConstants.LedCount));
