+++ src/main/java/frc/lib/BobcatLib/CANdle/CANdleState.java
@@ -2,29 +2,17 @@ package frc.lib.BobcatLib.CANdle;
 
 
 
-/*
 
-desired behavior, untested:
-
-intaking - fire animation while held
-intook - solid green 1 second
-intake stall - unbound
-reset pose - strobe gold 1 second
-reset gyro - strobe gold 1 second
-note hunting - rainbow while held
-outtake - strobe orange while held
+/**
+ * add season-specific states here
  */
-
 public enum CANdleState {
     OFF,
-    INTAKING,
-    INTOOK,
     INTAKESTALL,
     RESETPOSE,
-    NOTEHUNTING,
     RESETGYRO,
     OUTAKE,
-    FEED,
     ALIGNING,
     ALIGNED
+    
 }
