+++ src/main/java/frc/lib/Limelight/LimelightHelpersFast.java
@@ -1,6 +1,4 @@
-//LimelightHelpers v1.5.0 (March 27, 2024)
-
-package frc.lib.LimeLight;
+package frc.lib.Limelight;
 
 import edu.wpi.first.networktables.NetworkTable;
 import edu.wpi.first.networktables.NetworkTableEntry;
@@ -19,30 +17,27 @@ import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.concurrent.CompletableFuture;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.JsonFormat.Shape;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import com.alibaba.fastjson2.annotation.JSONField;
+import com.alibaba.fastjson.parser.*;
 
-public class LimelightHelpers {
+
+public class LimelightHelpersFast {
 
     public static class LimelightTarget_Retro {
 
-        @JsonProperty("t6c_ts")
+        @JSONField(name = "t6c_ts")
         private double[] cameraPose_TargetSpace;
 
-        @JsonProperty("t6r_fs")
+        @JSONField(name = "t6r_fs")
         private double[] robotPose_FieldSpace;
 
-        @JsonProperty("t6r_ts")
+        @JSONField(name = "t6r_ts")
         private  double[] robotPose_TargetSpace;
 
-        @JsonProperty("t6t_cs")
+        @JSONField(name = "t6t_cs")
         private double[] targetPose_CameraSpace;
 
-        @JsonProperty("t6t_rs")
+        @JSONField(name = "t6t_rs")
         private double[] targetPose_RobotSpace;
 
         public Pose3d getCameraPose_TargetSpace()
@@ -87,22 +82,22 @@ public class LimelightHelpers {
             return toPose2D(targetPose_RobotSpace);
         }
 
-        @JsonProperty("ta")
+        @JSONField(name = "ta")
         public double ta;
 
-        @JsonProperty("tx")
+        @JSONField(name = "tx")
         public double tx;
 
-        @JsonProperty("txp")
+        @JSONField(name = "txp")
         public double tx_pixels;
 
-        @JsonProperty("ty")
+        @JSONField(name = "ty")
         public double ty;
 
-        @JsonProperty("typ")
+        @JSONField(name = "typ")
         public double ty_pixels;
 
-        @JsonProperty("ts")
+        @JSONField(name = "ts")
         public double ts;
 
         public LimelightTarget_Retro() {
@@ -117,25 +112,25 @@ public class LimelightHelpers {
 
     public static class LimelightTarget_Fiducial {
 
-        @JsonProperty("fID")
+        @JSONField(name = "fID")
         public double fiducialID;
 
-        @JsonProperty("fam")
+        @JSONField(name = "fam")
         public String fiducialFamily;
 
-        @JsonProperty("t6c_ts")
+        @JSONField(name = "t6c_ts")
         private double[] cameraPose_TargetSpace;
 
-        @JsonProperty("t6r_fs")
+        @JSONField(name = "t6r_fs")
         private double[] robotPose_FieldSpace;
 
-        @JsonProperty("t6r_ts")
+        @JSONField(name = "t6r_ts")
         private double[] robotPose_TargetSpace;
 
-        @JsonProperty("t6t_cs")
+        @JSONField(name = "t6t_cs")
         private double[] targetPose_CameraSpace;
 
-        @JsonProperty("t6t_rs")
+        @JSONField(name = "t6t_rs")
         private double[] targetPose_RobotSpace;
 
         public Pose3d getCameraPose_TargetSpace()
@@ -180,22 +175,22 @@ public class LimelightHelpers {
             return toPose2D(targetPose_RobotSpace);
         }
         
-        @JsonProperty("ta")
+        @JSONField(name = "ta")
         public double ta;
 
-        @JsonProperty("tx")
+        @JSONField(name = "tx")
         public double tx;
 
-        @JsonProperty("txp")
+        @JSONField(name = "txp")
         public double tx_pixels;
 
-        @JsonProperty("ty")
+        @JSONField(name = "ty")
         public double ty;
 
-        @JsonProperty("typ")
+        @JSONField(name = "typ")
         public double ty_pixels;
 
-        @JsonProperty("ts")
+        @JSONField(name = "ts")
         public double ts;
         
         public LimelightTarget_Fiducial() {
@@ -213,28 +208,28 @@ public class LimelightHelpers {
 
     public static class LimelightTarget_Classifier {
 
-        @JsonProperty("class")
+        @JSONField(name = "class")
         public String className;
 
-        @JsonProperty("classID")
+        @JSONField(name = "classID")
         public double classID;
 
-        @JsonProperty("conf")
+        @JSONField(name = "conf")
         public double confidence;
 
-        @JsonProperty("zone")
+        @JSONField(name = "zone")
         public double zone;
 
-        @JsonProperty("tx")
+        @JSONField(name = "tx")
         public double tx;
 
-        @JsonProperty("txp")
+        @JSONField(name = "txp")
         public double tx_pixels;
 
-        @JsonProperty("ty")
+        @JSONField(name = "ty")
         public double ty;
 
-        @JsonProperty("typ")
+        @JSONField(name = "typ")
         public double ty_pixels;
 
         public  LimelightTarget_Classifier() {
@@ -243,28 +238,28 @@ public class LimelightHelpers {
 
     public static class LimelightTarget_Detector {
 
-        @JsonProperty("class")
+        @JSONField(name = "class")
         public String className;
 
-        @JsonProperty("classID")
+        @JSONField(name = "classID")
         public double classID;
 
-        @JsonProperty("conf")
+        @JSONField(name = "conf")
         public double confidence;
 
-        @JsonProperty("ta")
+        @JSONField(name = "ta")
         public double ta;
 
-        @JsonProperty("tx")
+        @JSONField(name = "tx")
         public double tx;
 
-        @JsonProperty("txp")
+        @JSONField(name = "txp")
         public double tx_pixels;
 
-        @JsonProperty("ty")
+        @JSONField(name = "ty")
         public double ty;
 
-        @JsonProperty("typ")
+        @JSONField(name = "typ")
         public double ty_pixels;
 
         public LimelightTarget_Detector() {
@@ -273,49 +268,33 @@ public class LimelightHelpers {
 
     public static class Results {
 
-        @JsonProperty("pID")
+        @JSONField(name = "pID")
         public double pipelineID;
 
-        @JsonProperty("tl")
+        @JSONField(name = "tl")
         public double latency_pipeline;
 
-        @JsonProperty("cl")
+        @JSONField(name = "cl")
         public double latency_capture;
 
         public double latency_jsonParse;
 
-        @JsonProperty("ts")
+        @JSONField(name = "ts")
         public double timestamp_LIMELIGHT_publish;
 
-        @JsonProperty("ts_rio")
+        @JSONField(name = "ts_rio")
         public double timestamp_RIOFPGA_capture;
 
-        @JsonProperty("v")
-        @JsonFormat(shape = Shape.NUMBER)
-        public boolean valid;
-
-        @JsonProperty("botpose")
+        @JSONField(name = "botpose")
         public double[] botpose;
 
-        @JsonProperty("botpose_wpired")
+        @JSONField(name = "botpose_wpired")
         public double[] botpose_wpired;
 
-        @JsonProperty("botpose_wpiblue")
+        @JSONField(name = "botpose_wpiblue")
         public double[] botpose_wpiblue;
 
-        @JsonProperty("botpose_tagcount")
-        public double botpose_tagcount;
-       
-        @JsonProperty("botpose_span")
-        public double botpose_span;
-       
-        @JsonProperty("botpose_avgdist")
-        public double botpose_avgdist;
-       
-        @JsonProperty("botpose_avgarea")
-        public double botpose_avgarea;
-
-        @JsonProperty("t6c_rs")
+        @JSONField(name = "t6c_rs")
         public double[] camerapose_robotspace;
 
         public Pose3d getBotPose3d() {
@@ -342,19 +321,19 @@ public class LimelightHelpers {
             return toPose2D(botpose_wpiblue);
         }
 
-        @JsonProperty("Retro")
+        @JSONField(name = "Retro")
         public LimelightTarget_Retro[] targets_Retro;
 
-        @JsonProperty("Fiducial")
+        @JSONField(name = "Fiducial")
         public LimelightTarget_Fiducial[] targets_Fiducials;
 
-        @JsonProperty("Classifier")
+        @JSONField(name = "Classifier")
         public LimelightTarget_Classifier[] targets_Classifier;
 
-        @JsonProperty("Detector")
+        @JSONField(name = "Detector")
         public LimelightTarget_Detector[] targets_Detector;
 
-        @JsonProperty("Barcode")
+        @JSONField(name = "Barcode")
         public LimelightTarget_Barcode[] targets_Barcode;
 
         public Results() {
@@ -372,66 +351,14 @@ public class LimelightHelpers {
     }
 
     public static class LimelightResults {
-        @JsonProperty("Results")
+        @JSONField(name = "Results")
         public Results targetingResults;
-        
-        public String error;
 
         public LimelightResults() {
             targetingResults = new Results();
-            error = "";
-        }
-
-
-    }
-
-    public static class RawFiducial {
-        public int id;
-        public double txnc;
-        public double tync;
-        public double ta;
-        public double distToCamera;
-        public double distToRobot;
-        public double ambiguity;
-
-
-        public RawFiducial(int id, double txnc, double tync, double ta, double distToCamera, double distToRobot, double ambiguity) {
-            this.id = id;
-            this.txnc = txnc;
-            this.tync = tync;
-            this.ta = ta;
-            this.distToCamera = distToCamera;
-            this.distToRobot = distToRobot;
-            this.ambiguity = ambiguity;
         }
     }
 
-    public static class PoseEstimate {
-        public Pose2d pose;
-        public double timestampSeconds;
-        public double latency;
-        public int tagCount;
-        public double tagSpan;
-        public double avgTagDist;
-        public double avgTagArea;
-        public RawFiducial[] rawFiducials; 
-
-        public PoseEstimate(Pose2d pose, double timestampSeconds, double latency, 
-            int tagCount, double tagSpan, double avgTagDist, 
-            double avgTagArea, RawFiducial[] rawFiducials) {
-
-            this.pose = pose;
-            this.timestampSeconds = timestampSeconds;
-            this.latency = latency;
-            this.tagCount = tagCount;
-            this.tagSpan = tagSpan;
-            this.avgTagDist = avgTagDist;
-            this.avgTagArea = avgTagArea;
-            this.rawFiducials = rawFiducials;
-        }
-    }
-
-    private static ObjectMapper mapper;
 
     /**
      * Print JSON Parse time to the console in milliseconds
@@ -466,86 +393,6 @@ public class LimelightHelpers {
         Translation2d tran2d = new Translation2d(inData[0], inData[1]);
         Rotation2d r2d = new Rotation2d(Units.degreesToRadians(inData[5]));
         return new Pose2d(tran2d, r2d);
-    }
-
-    private static double extractBotPoseEntry(double[] inData, int position){
-        if(inData.length < position+1)
-        {
-            return 0;
-        }
-        return inData[position];
-    }
-
-    private static PoseEstimate getBotPoseEstimate(String limelightName, String entryName) {
-        var poseEntry = LimelightHelpers.getLimelightNTTableEntry(limelightName, entryName);
-        var poseArray = poseEntry.getDoubleArray(new double[0]);
-        var pose = toPose2D(poseArray);
-        double latency = extractBotPoseEntry(poseArray,6);
-        int tagCount = (int)extractBotPoseEntry(poseArray,7);
-        double tagSpan = extractBotPoseEntry(poseArray,8);
-        double tagDist = extractBotPoseEntry(poseArray,9);
-        double tagArea = extractBotPoseEntry(poseArray,10);
-        //getlastchange() in microseconds, ll latency in milliseconds
-        var timestamp = (poseEntry.getLastChange() / 1000000.0) - (latency/1000.0);
-
-
-        RawFiducial[] rawFiducials = new RawFiducial[tagCount];
-        int valsPerFiducial = 7;
-        int expectedTotalVals = 11 + valsPerFiducial*tagCount;
-
-        if (poseArray.length != expectedTotalVals) {
-            // Don't populate fiducials
-        }
-        else{
-            for(int i = 0; i < tagCount; i++) {
-                int baseIndex = 11 + (i * valsPerFiducial);
-                int id = (int)poseArray[baseIndex];
-                double txnc = poseArray[baseIndex + 1];
-                double tync = poseArray[baseIndex + 2];
-                double ta = poseArray[baseIndex + 3];
-                double distToCamera = poseArray[baseIndex + 4];
-                double distToRobot = poseArray[baseIndex + 5];
-                double ambiguity = poseArray[baseIndex + 6];
-                rawFiducials[i] = new RawFiducial(id, txnc, tync, ta, distToCamera, distToRobot, ambiguity);
-            }
-        }
-
-        return new PoseEstimate(pose, timestamp,latency,tagCount,tagSpan,tagDist,tagArea,rawFiducials);
-    }
-
-    private static void printPoseEstimate(PoseEstimate pose) {
-        if (pose == null) {
-            System.out.println("No PoseEstimate available.");
-            return;
-        }
-    
-        System.out.printf("Pose Estimate Information:%n");
-        System.out.printf("Timestamp (Seconds): %.3f%n", pose.timestampSeconds);
-        System.out.printf("Latency: %.3f ms%n", pose.latency);
-        System.out.printf("Tag Count: %d%n", pose.tagCount);
-        System.out.printf("Tag Span: %.2f meters%n", pose.tagSpan);
-        System.out.printf("Average Tag Distance: %.2f meters%n", pose.avgTagDist);
-        System.out.printf("Average Tag Area: %.2f%% of image%n", pose.avgTagArea);
-        System.out.println();
-    
-        if (pose.rawFiducials == null || pose.rawFiducials.length == 0) {
-            System.out.println("No RawFiducials data available.");
-            return;
-        }
-    
-        System.out.println("Raw Fiducials Details:");
-        for (int i = 0; i < pose.rawFiducials.length; i++) {
-            RawFiducial fiducial = pose.rawFiducials[i];
-            System.out.printf(" Fiducial #%d:%n", i + 1);
-            System.out.printf("  ID: %d%n", fiducial.id);
-            System.out.printf("  TXNC: %.2f%n", fiducial.txnc);
-            System.out.printf("  TYNC: %.2f%n", fiducial.tync);
-            System.out.printf("  TA: %.2f%n", fiducial.ta);
-            System.out.printf("  Distance to Camera: %.2f meters%n", fiducial.distToCamera);
-            System.out.printf("  Distance to Robot: %.2f meters%n", fiducial.distToRobot);
-            System.out.printf("  Ambiguity: %.2f%n", fiducial.ambiguity);
-            System.out.println();
-        }
     }
 
     public static NetworkTable getLimelightNTTable(String tableName) {
@@ -583,7 +430,7 @@ public class LimelightHelpers {
             url = new URL(urlString);
             return url;
         } catch (MalformedURLException e) {
-            System.err.println("bad LL URL");
+            //System.err.println("bad LL URL");
         }
         return null;
     }
@@ -687,8 +534,8 @@ public class LimelightHelpers {
         return getLimelightNTDouble(limelightName, "tid");
     }
 
-    public static String getNeuralClassID(String limelightName) {
-        return getLimelightNTString(limelightName, "tclass");
+    public static double getNeuralClassID(String limelightName) {
+        return getLimelightNTDouble(limelightName, "tclass");
     }
 
     /////
@@ -748,28 +595,6 @@ public class LimelightHelpers {
     }
 
     /**
-     * Gets the Pose2d and timestamp for use with WPILib pose estimator (addVisionMeasurement) when you are on the BLUE
-     * alliance
-     * 
-     * @param limelightName
-     * @return
-     */
-    public static PoseEstimate getBotPoseEstimate_wpiBlue(String limelightName) {
-        return getBotPoseEstimate(limelightName, "botpose_wpiblue");
-    }
-
-    /**
-     * Gets the Pose2d and timestamp for use with WPILib pose estimator (addVisionMeasurement) when you are on the BLUE
-     * alliance
-     * 
-     * @param limelightName
-     * @return
-     */
-    public static PoseEstimate getBotPoseEstimate_wpiBlue_MegaTag2(String limelightName) {
-        return getBotPoseEstimate(limelightName, "botpose_orb_wpiblue");
-    }
-
-    /**
      * Gets the Pose2d for easy use with Odometry vision pose estimator
      * (addVisionMeasurement)
      * 
@@ -781,26 +606,6 @@ public class LimelightHelpers {
         double[] result = getBotPose_wpiRed(limelightName);
         return toPose2D(result);
 
-    }
-
-    /**
-     * Gets the Pose2d and timestamp for use with WPILib pose estimator (addVisionMeasurement) when you are on the RED
-     * alliance
-     * @param limelightName
-     * @return
-     */
-    public static PoseEstimate getBotPoseEstimate_wpiRed(String limelightName) {
-        return getBotPoseEstimate(limelightName, "botpose_wpired");
-    }
-
-    /**
-     * Gets the Pose2d and timestamp for use with WPILib pose estimator (addVisionMeasurement) when you are on the RED
-     * alliance
-     * @param limelightName
-     * @return
-     */
-    public static PoseEstimate getBotPoseEstimate_wpiRed_MegaTag2(String limelightName) {
-        return getBotPoseEstimate(limelightName, "botpose_orb_wpired");
     }
 
     /**
@@ -826,11 +631,6 @@ public class LimelightHelpers {
 
     public static void setPipelineIndex(String limelightName, int pipelineIndex) {
         setLimelightNTDouble(limelightName, "pipeline", pipelineIndex);
-    }
-
-    
-    public static void setPriorityTagID(String limelightName, int ID) {
-        setLimelightNTDouble(limelightName, "priorityid", ID);
     }
 
     /**
@@ -886,28 +686,6 @@ public class LimelightHelpers {
         setLimelightNTDoubleArray(limelightName, "crop", entries);
     }
 
-    public static void SetRobotOrientation(String limelightName, double yaw, double yawRate, 
-        double pitch, double pitchRate, 
-        double roll, double rollRate) {
-
-        double[] entries = new double[6];
-        entries[0] = yaw;
-        entries[1] = yawRate;
-        entries[2] = pitch;
-        entries[3] = pitchRate;
-        entries[4] = roll;
-        entries[5] = rollRate;
-        setLimelightNTDoubleArray(limelightName, "robot_orientation_set", entries);
-    }
-
-    public static void SetFiducialIDFiltersOverride(String limelightName, int[] validIDs) {
-        double[] validIDsDouble = new double[validIDs.length];
-        for (int i = 0; i < validIDs.length; i++) {
-            validIDsDouble[i] = validIDs[i];
-        }        
-        setLimelightNTDoubleArray(limelightName, "fiducial_id_filters_set", validIDsDouble);
-    }
-    
     public static void setCameraPose_RobotSpace(String limelightName, double forward, double side, double up, double roll, double pitch, double yaw) {
         double[] entries = new double[6];
         entries[0] = forward;
@@ -955,7 +733,7 @@ public class LimelightHelpers {
             if (responseCode == 200) {
                 return true;
             } else {
-                System.err.println("Bad LL Request");
+                //System.err.println("Bad LL Request");
             }
         } catch (IOException e) {
             System.err.println(e.getMessage());
@@ -966,27 +744,37 @@ public class LimelightHelpers {
     /**
      * Parses Limelight's JSON results dump into a LimelightResults Object
      */
-    public static LimelightResults getLatestResults(String limelightName) {
+// ...
 
-        long start = System.nanoTime();
-        LimelightHelpers.LimelightResults results = new LimelightHelpers.LimelightResults();
-        if (mapper == null) {
-            mapper = new ObjectMapper().configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-        }
+public static LimelightResults getLatestResults(String limelightName) {
+    long start = System.nanoTime();
+    LimelightHelpersFast.LimelightResults results = new LimelightHelpersFast.LimelightResults();
 
-        try {
-            results = mapper.readValue(getJSONDump(limelightName), LimelightResults.class);
-        } catch (JsonProcessingException e) {
-            results.error = "lljson error: " + e.getMessage();
-        }
+    try {
+        String jsonDump = getJSONDump(limelightName);
 
-        long end = System.nanoTime();
-        double millis = (end - start) * .000001;
-        results.targetingResults.latency_jsonParse = millis;
-        if (profileJSON) {
-            System.out.printf("lljson: %.2f\r\n", millis);
-        }
+        // Configure parser to ignore unknown properties
+        ParserConfig config = new ParserConfig();
 
-        return results;
+        // Use DefaultJSONParser to parse with configuration
+        DefaultJSONParser jsonParser = new DefaultJSONParser(jsonDump, config);
+        results = jsonParser.parseObject(LimelightResults.class);
+    } catch (Exception e) {
+        System.err.println("lljson error: " + e.getMessage());
     }
+
+    long end = System.nanoTime();
+    double millis = (end - start) * .000001;
+    results.targetingResults.latency_jsonParse = millis;
+    if (profileJSON) {
+        System.out.printf("lljson: %.2f\r\n", millis);
+    }
+
+    return results;
 }
+
+    
+
+
+
+    }
